remove media component
