+++ src/core/client/admin/components/MediaContainer/MediaContainer.tsx
@@ -5,7 +5,10 @@ import { withFragmentContainer } from "coral-framework/lib/relay";
 
 import { MediaContainer_comment } from "coral-admin/__generated__/MediaContainer_comment.graphql";
 
-import Media from "./Media";
+import ExternalMedia from "./ExternalMedia";
+import GiphyMedia from "./GiphyMedia";
+import TwitterMedia from "./TwitterMedia";
+import YouTubeMedia from "./YouTubeMedia";
 
 interface Props {
   comment: MediaContainer_comment;
@@ -17,18 +20,46 @@ const MediaContainer: FunctionComponent<Props> = ({ comment }) => {
     return null;
   }
 
-  if (media.__typename === "%other") {
-    return null;
+  switch (media.__typename) {
+    case "GiphyMedia":
+      return (
+        <GiphyMedia
+          still={media.still}
+          video={media.video}
+          title={media.title}
+          width={media.width}
+          height={media.height}
+        />
+      );
+    case "ExternalMedia":
+      return (
+        <ExternalMedia
+          id={comment.id}
+          url={media.url}
+          siteID={comment.site.id}
+        />
+      );
+    case "TwitterMedia":
+      return (
+        <TwitterMedia
+          id={comment.id}
+          url={media.url}
+          siteID={comment.site.id}
+        />
+      );
+    case "YouTubeMedia":
+      return (
+        <YouTubeMedia
+          id={comment.id}
+          url={media.url}
+          siteID={comment.site.id}
+          still={media.still}
+          title={media.title}
+        />
+      );
+    case "%other":
+      return null;
   }
-
-  return (
-    <Media
-      media={media}
-      type={media.__typename}
-      id={comment.id}
-      siteID={comment.site.id}
-    />
-  );
 };
 
 const enhanced = withFragmentContainer<Props>({
