+++ src/core/client/admin/components/MediaContainer/TwitterMedia.tsx
@@ -4,7 +4,7 @@ import Frame from "coral-framework/components/Frame";
 
 import styles from "./Media.css";
 
-export interface Props {
+interface Props {
   id: string;
   url: string;
   siteID: string;
