+++ src/core/client/admin/components/MediaContainer/YouTubeMedia.tsx
@@ -6,7 +6,7 @@ import { BaseButton, Flex, Icon } from "coral-ui/components/v2";
 
 import styles from "./Media.css";
 
-export interface Props {
+interface Props {
   id: string;
   url: string;
   still: string;
