+++ src/core/client/admin/components/ModerateCard/CommentRevisionContainer.tsx
@@ -3,7 +3,10 @@ import { graphql } from "react-relay";
 
 import { withFragmentContainer } from "coral-framework/lib/relay";
 import { HorizontalGutter, Timestamp } from "coral-ui/components/v2";
-import Media from "../MediaContainer/Media";
+import ExternalMedia from "../MediaContainer/ExternalMedia";
+import GiphyMedia from "../MediaContainer/GiphyMedia";
+import TwitterMedia from "../MediaContainer/TwitterMedia";
+import YouTubeMedia from "../MediaContainer/YouTubeMedia";
 
 import { CommentRevisionContainer_comment as CommentData } from "coral-admin/__generated__/CommentRevisionContainer_comment.graphql";
 
@@ -28,12 +31,36 @@ const CommentRevisionContainer: FunctionComponent<Props> = ({ comment }) => {
           <div key={c.id}>
             <Timestamp>{c.createdAt}</Timestamp>
             <CommentContent>{c.body ? c.body : ""}</CommentContent>
-            {c.media && c.media.__typename !== "%other" && (
-              <Media
-                media={c.media}
+            {c.media && c.media.__typename === "GiphyMedia" && (
+              <GiphyMedia
+                still={c.media.still}
+                video={c.media.video}
+                title={c.media.title}
+                width={c.media.width}
+                height={c.media.height}
+              />
+            )}
+            {c.media && c.media.__typename === "ExternalMedia" && (
+              <ExternalMedia
                 id={comment.id}
+                url={c.media.url}
                 siteID={comment.site.id}
-                type={c.media.__typename}
+              />
+            )}
+            {c.media && c.media.__typename === "TwitterMedia" && (
+              <TwitterMedia
+                id={comment.id}
+                url={c.media.url}
+                siteID={comment.site.id}
+              />
+            )}
+            {c.media && c.media.__typename === "YouTubeMedia" && (
+              <YouTubeMedia
+                id={comment.id}
+                url={c.media.url}
+                siteID={comment.site.id}
+                still={c.media.still}
+                title={c.media.title}
               />
             )}
           </div>
